move modal of Cheatsheet to Editor component
